streamlit_requirements.txt
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
streamlit==1.34.0
sqlite-vec==0.1.6
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.3
scikit-image==0.22.0
scikit-learn==1.3.0
opencv-python==4.8.0.74
